Move vectorization decorators to own file, use partial in core.py
