+++ tests/unit/test_core.py
@@ -39,88 +39,39 @@ class TestSimulatorOperation():
 
 
 class Test_vectorization():
-    """Test operation vectorization operations 'as_vectorized_*'
+    """Test operation vectorization
     """
-
-    @staticmethod
-    @elfi.core.as_vectorized_simulator
-    def mock_simulator1(*data, prng=None):
-        return np.array((1,))
-
-    @staticmethod
-    def mock_simulator2(*data, n_sim=1, prng=None):
-        return np.ones((n_sim, 1))
-
-    @staticmethod
-    @elfi.core.as_vectorized_summary
-    def mock_summary1(*data):
-        return np.ones((1,))
-
-    @staticmethod
-    def mock_summary2(*data):
-        n_sim = data[0].shape[0]
-        return np.ones((n_sim, 1))
-
-    @staticmethod
-    @elfi.core.as_vectorized_discrepancy
-    def mock_discrepancy1(x, y):
-        return np.ones((1,))
-
-    @staticmethod
-    def mock_discrepancy2(x, y):
-        n_sim = x[0].shape[0]
-        return np.ones((n_sim, 1))
 
     def test_as_vectorized_simulator(self):
         ret1 = np.array([5])
         ret2 = np.array([6])
-        mock_seq = elfi.core.as_vectorized_simulator(MockSequentialSimulator([ret1, ret2]))
+        mock_seq = partial(elfi.core.vectorize_simulator, MockSequentialSimulator([ret1, ret2]))
         mock_vec = MockSimulator([ret1, ret2])
         input_data = [np.atleast_2d([[1], [2]]), np.atleast_2d([[3], [4]])]
         output_seq = mock_seq(*input_data, n_sim=2)
         output_vec = mock_vec(*input_data, n_sim=2)
         assert np.array_equal(output_seq, output_vec)
 
-    def test_annotated_simulator(self):
-        input_data = [np.atleast_2d([[1], [2]]), np.atleast_2d([[3], [4]])]
-        output1 = self.mock_simulator1(*input_data, n_sim=2)
-        output2 = self.mock_simulator2(*input_data, n_sim=2)
-        assert np.array_equal(output1, output2)
-
     def test_as_vectorized_summary(self):
         ret1 = np.array([5])
         ret2 = np.array([6])
-        mock_seq = elfi.core.as_vectorized_summary(MockSequentialSummary([ret1, ret2]))
+        mock_seq = partial(elfi.core.vectorize_summary, MockSequentialSummary([ret1, ret2]))
         mock_vec = MockSummary([ret1, ret2])
         input_data = [np.atleast_2d([[1], [2]]), np.atleast_2d([[3], [4]])]
         output_seq = mock_seq(*input_data)
         output_vec = mock_vec(*input_data)
         assert np.array_equal(output_seq, output_vec)
 
-    def test_annotated_summary(self):
-        input_data = [np.atleast_2d([[1], [2]]), np.atleast_2d([[3], [4]])]
-        output1 = self.mock_summary1(*input_data)
-        output2 = self.mock_summary2(*input_data)
-        assert np.array_equal(output1, output2)
-
     def test_as_vectorized_discrepancy(self):
         ret1 = np.array([5])
         ret2 = np.array([6])
-        mock_seq = elfi.core.as_vectorized_discrepancy(MockSequentialDiscrepancy([ret1, ret2]))
+        mock_seq = partial(elfi.core.vectorize_discrepancy, MockSequentialDiscrepancy([ret1, ret2]))
         mock_vec = MockDiscrepancy([ret1, ret2])
         x = (np.atleast_2d([[1], [2]]), np.atleast_2d([[3], [4]]))
         y = (np.atleast_2d([[5]]), np.atleast_2d([[6]]))
         output_seq = mock_seq(x, y)
         output_vec = mock_vec(x, y)
         assert np.array_equal(output_seq, output_vec)
-
-    def test_annotated_discrepancy(self):
-        x = (np.atleast_2d([[1], [2]]), np.atleast_2d([[3], [4]]))
-        y = (np.atleast_2d([[5]]), np.atleast_2d([[6]]))
-        output1 = self.mock_discrepancy1(x, y)
-        output2 = self.mock_discrepancy2(x, y)
-        assert np.array_equal(output1, output2)
-
 
 
 def test_node_data_sub_slicing():
